♻ refactor: actions 이름 수정
